cleanup(libscap): create dedicated scap_machine_agent
